Chatbot_DSA/requirements.txt
streamlit =1.11.1
altair
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai